python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "totalreclaw"
version = "0.1.0"
description = "End-to-end encrypted memory for AI agents — Python client"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "TotalReclaw Team" }]
dependencies = [
    "cryptography>=42.0",
    "mnemonic>=0.21",
    "httpx>=0.27",
    "onnxruntime>=1.17",
    "numpy>=1.26",
    "PyStemmer>=2.2",
    "transformers>=4.40",
    "huggingface-hub>=0.23",
    "eth-account>=0.13",
]

[project.optional-dependencies]
hermes = ["hermes-agent>=0.5.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "pytest-xdist>=3.5"]

[project.entry-points."hermes_agent.plugins"]
totalreclaw = "totalreclaw.hermes"

[tool.pytest.ini_options]
asyncio_mode = "auto"
